카테고리 제거 시 널값 체크 로직 추가
카테고리를 제거하는 로직에 널 값을 체크해서 예외를 던지는 로직 추가

+++ src/main/java/com/mylog/post/service/CategoryService.java
@@ -50,6 +50,10 @@ public class CategoryService {
 
     public void removeCategory(Long id) {
         Category category = categoryRepository.findById(id).orElse(null);
+        System.out.println("category = " + category);
+        if (category == null) {
+            throw new IllegalStateException("해당 카테고리가 존재하지 않습니다.");
+        }
 
         categoryRepository.delete(category);
     }
